python-flask-gunicorn-nginx/{{ cookiecutter.project_slug }}/requirements.txt
# Python dependencies
Flask
python-dotenv